BCW-00003 Поднял mysql контейнер в докере, организовал сетевую связь с веб контейнером и исправил мелкие замечания
