+++ README.md
@@ -4,4 +4,6 @@ Botec web interface
 Для запуска проекта - php artisan serve
 Зависимости нужно через composer подтянуть: composer install  & composer update  (нужно перейти в корневую папку botec-web)
 Для запуска докер нужно перейти в дир: /build/dev
-Для запуска докер: docker-compose up 
+Для запуска докер: docker-compose up
+
+Для проверки статуса БД, переходим по ссылке: http://127.0.0.1:8080/database
