+++ config/database.php
@@ -46,11 +46,11 @@ return [
         'mysql' => [
             'driver' => 'mysql',
             'url' => env('DATABASE_URL'),
-            'host' => env('DB_HOST', 'db'),
-            'port' => env('DB_PORT', '3306'),
-            'database' => env('DB_DATABASE', 'botec-web'),
-            'username' => env('DB_USERNAME', 'botec-web'),
-            'password' => env('DB_PASSWORD', 'botec-web'),
+            'host' => env('DB_HOST', ''),
+            'port' => env('DB_PORT', ''),
+            'database' => env('DB_DATABASE', ''),
+            'username' => env('DB_USERNAME', ''),
+            'password' => env('DB_PASSWORD', ''),
             'unix_socket' => env('DB_SOCKET', ''),
             'charset' => 'utf8mb4',
             'collation' => 'utf8mb4_unicode_ci',
